sos muy bueno

+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/models/Resultado.java
@@ -13,8 +13,8 @@ public class Resultado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    int orden;
+    Long orden;
     String nombre;
-    int votos;
-    double porcentaje;
+    Integer votos;
+    Double porcentaje;
 }
